requirements.txt
openai>=1.0.0
jinja2>=3.1.2
textstat>=0.7.3
python-dotenv>=1.0.1
